fix sector offset calculation for media with 256 sectors per track DFI listing/loading now works cleanup unneeded cpp files

+++ lib/meatloaf/disk/d64.h
@@ -76,7 +76,7 @@ protected:
 
 public:
     std::vector<Partition> partitions;
-    std::vector<uint8_t> sectorsPerTrack = { 17, 18, 19, 21 };
+    std::vector<uint16_t> sectorsPerTrack = { 17, 18, 19, 21 };
     std::vector<uint8_t> interleave = { 3, 10 }; // Directory, File
 
     uint8_t dos_version = 0x41;
